+++ agentveil/__init__.py
@@ -24,7 +24,7 @@ from agentveil.exceptions import (
     AVPServerError,
 )
 
-__version__ = "0.5.8"
+__version__ = "0.6.1"
 
 __all__ = [
     "AVPAgent",